admin role

+++ back/controllers/messages.js
@@ -1,4 +1,5 @@
 const Message = require("../models/Message");
+const User = require("../models/User");
 
 exports.createMessage = (req, res, next) => {
   const messageObject = req.body;
@@ -23,19 +24,22 @@ exports.getAllMessages = (req, res, next) => {
     .catch((error) => res.status(404).json(error));
 };
 exports.deleteMessage = (req, res, next) => {
-  Message.findOne({ _id: req.params.id })
-    .then((message) => {
-      if (message.userId != req.auth.userId) {
-        res.status(401).json({ message: "Not authorized" });
-      } else {
-        Message.deleteOne({ _id: req.params.id })
-          .then((infos) => {
-            res.status(200).json(req.params.id);
-          })
-          .catch((error) => res.status(401).json({ error }));
-      }
-    })
-    .catch((error) => {
-      res.status(500).json({ error });
-    });
+  user = User.findOne({ _id: req.auth.userId }).then((user) => {
+    const user_role = user.role;
+    Message.findOne({ _id: req.params.id })
+      .then((message) => {
+        if (message.userId != req.auth.userId && user_role != "admin") {
+          res.status(401).json({ message: "Not authorized" });
+        } else {
+          Message.deleteOne({ _id: req.params.id })
+            .then((infos) => {
+              res.status(200).json(req.params.id);
+            })
+            .catch((error) => res.status(401).json({ error }));
+        }
+      })
+      .catch((error) => {
+        res.status(500).json({ error });
+      });
+  });
 };
